Edit Blocks

These functions are assigned to blocks inside plan_sec(b* = fun(...)). See planning. e and div edit existing blocks, the rest adds new markdown content.

e is used on single element blocks ('text expressions'). It splits the expression into clauses, and allows the user to select and alter each.

div is used on multi-element blocks. Currently, it returns the block as is.

add_cur_head adds the heading of the current section, with the specified level.

Usage

e(cs = NULL, subs = NULL, pat = NULL)

div()

add_cur_head(n = NULL)

Arguments

cs

Integer vector of clauses to keep. Defaults to all.

subs

Character vector specifying substitutions for the final punctuation on each clause in cs. Defaults to none.

pat

The pattern to break expressions into clauses. Defaults to "(?<=[\\.:;!?])".

n

The heading level to use.