Find build information of a ggplot

To customize a plot made in a package, adding new layers, it is useful to see the data and mappings used. This function returns a list with both.

Usage

get_gg_info(graph)

Arguments

graph

A ggplot to analyze.

Value

A list with two entries:

main_data: the data used in ggplot().

layers: for each layer the aesthetic mapping and the data used (which can be 'inherited').

Examples

g <- ggplot2::ggplot(mtcars, ggplot2::aes(mpg, hp)) +
  ggplot2::geom_point(ggplot2::aes(color = gear))
get_gg_info(g)
#> $main_data
#> # A tibble: 32 × 11
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # ℹ 22 more rows
#> 
#> $layers
#> $layers$geom_point
#> $layers$geom_point$mappings
#> NULL
#> 
#> $layers$geom_point$data
#> [1] "inherited"
#> 
#> 
#>