Plan Document Morphing

These functions are used to programmatically alter the Rmd, once it has been split (see splitting).

In plan_doc, global changes can be defined. Then, the user must supply one s* = plan_sec(...) arguments, where * corresponds to the section number, for each section he wants to keep.

In plan_sec, local section changes can be defined. Then, the user must supply one b* = fun(...) to each block he wants to keep. There is also the possibility to add new lines of markdown with unnamed calls. See the available block editing functions in editing.

Usage

plan_doc(sections, level_add = 0, ...)

plan_sec(head_level = NULL, ...)

Arguments

sections

A list of sections as returned by split_sections.

level_add

An integer value to adjust the header level of all sections. For instance, a level_add = 1 transform "# Header" into "## Header". Negative values decrease the header value.

...

s* = plan_sec(...) arguments for plan_doc, or b* = fun(...) arguments for plan_sec.

head_level

An integer value to set the heading level of the current section header.

Value

For plan_doc, a character vector containing the processed document. plan_sec should not be used outside plan_doc.