Split Rmd Into Sections and Blocks

This function takes a path to an R Markdown (.Rmd) file or a string vector of text from an Rmd file, cleans it by replacing specified patterns, and then splits it into sections and blocks. The result makes it easy to refer to each section/block with the plan functions of the package.

Usage

split_sections(x, str_pattern = NULL, split_sec_lim = 3, split_args = NULL)

Arguments

x

A character vector containing the path to an Rmd file or the actual text from an Rmd file.

str_pattern

A named vector where names are patterns to match in the document's text and values are the replacements for those patterns. Passed to the pattern argument of str_replace_all. Defaults to no substitution.

split_sec_lim

An integer specifying the maximum heading size (number) of "#" to be considered. Sub-sections lower than that won't be split.

split_args

A list of 'splits' to separate the sections and blocks. See the source code of morphdown:::apply_splits for the default splits, and morphdown:::split_match to see how they work.

Value

A named list of character vectors. Each vector is a section named s*. Each section has blocks named b*. Use those names to refer to each in the plan functions.