varr: Utility Functions for VAR Models

Utility functions for VAR models. Currently, ggplots for objects like like IRF's, FEVD's, predict graphs, and residual analysis. Soon to contain functions for common tests and statistics, support for other VAR-like objects, and unified builders for vector regression models.

See also

Author

Maintainer: Ricardo Semião ricardo.semiao@outlook.com